+++ changelog.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.0-alpha.5 - 2020-10-01
+## 1.0.0-alpha.5 - 2020-10-05
 
 ### Added
 
--   Added WP-Cron substitute in case of sites with modified wp installation
+-   Added WP-Cron substitute in case of sites with modified wp installation.
+-   Added self-hosted plugin updates.
 
 ## 1.0.0-alpha.4 - 2020-09-25